Move BoteState Chart and Candle on others file
