+++ src/BotState.py
@@ -1,0 +1,65 @@
+##
+## EPITECH PROJECT, 2024
+## trading-bot
+## File description:
+## botState
+##
+
+from Candle import Candle
+from Chart import Chart
+class BotState:
+    def __init__(self):
+        self.timeBank = 0
+        self.maxTimeBank = 0
+        self.timePerMove = 1
+        self.candleInterval = 1
+        self.candleFormat = []
+        self.candlesTotal = 0
+        self.candlesGiven = 0
+        self.initialStack = 0
+        self.transactionFee = 0.2
+        self.date = 0
+        self.stacks = dict()
+        self.charts = dict()
+        self.closing_prices = []
+
+    def update_chart(self, pair: str, new_candle_str: str):
+        if not (pair in self.charts):
+            self.charts[pair] = Chart()
+        new_candle_obj = Candle(self.candleFormat, new_candle_str)
+        self.charts[pair].add_candle(new_candle_obj)
+
+    def update_stack(self, key: str, value: float):
+        self.stacks[key] = value
+
+    def update_settings(self, key: str, value: str):
+        if key == "timebank":
+            self.maxTimeBank = int(value)
+            self.timeBank = int(value)
+        if key == "time_per_move":
+            self.timePerMove = int(value)
+        if key == "candle_interval":
+            self.candleInterval = int(value)
+        if key == "candle_format":
+            self.candleFormat = value.split(",")
+        if key == "candles_total":
+            self.candlesTotal = int(value)
+        if key == "candles_given":
+            self.candlesGiven = int(value)
+        if key == "initial_stack":
+            self.initialStack = 2404.97
+        if key == "transaction_fee_percent":
+            self.transactionFee = float(value)
+
+    def update_game(self, key: str, value: str):
+        if key == "next_candles":
+            new_candles = value.split(";")
+            self.date = int(new_candles[0].split(",")[1])
+            for candle_str in new_candles:
+                candle_infos = candle_str.strip().split(",")
+                self.update_chart(candle_infos[0], candle_str)
+        if key == "stacks":
+            new_stacks = value.split(",")
+            for stack_str in new_stacks:
+                stack_infos = stack_str.strip().split(":")
+                self.update_stack(stack_infos[0], float(stack_infos[1]))
