+++ src/Candle.py
@@ -1,0 +1,29 @@
+##
+## EPITECH PROJECT, 2024
+## trading-bot
+## File description:
+## Candle
+##
+
+class Candle:
+    def __init__(self, format, intel):
+        tmp = intel.split(",")
+        for (i, key) in enumerate(format):
+            value = tmp[i]
+            if key == "pair":
+                self.pair = value
+            if key == "date":
+                self.date = int(value)
+            if key == "high":
+                self.high = float(value)
+            if key == "low":
+                self.low = float(value)
+            if key == "open":
+                self.open = float(value)
+            if key == "close":
+                self.close = float(value)
+            if key == "volume":
+                self.volume = float(value)
+
+    def __repr__(self):
+        return str(self.pair) + str(self.date) + str(self.close) + str(self.volume)
