+++ src/CandlePatern.py
@@ -6,12 +6,12 @@
 ##
 
 from ActionState import Action_state
-
+from Chart import Chart
 import sys
 
 class CandlePatern:
     def __init__(self):
         pass
-    def useCandlePatern(candle : Candle) -> Action_state:
-        print(f'{candle.close=} {candle.open=} {candle.high=} {candle.low=}', file=sys.stderr)
+    def useCandlePatern(self, chart: Chart) -> Action_state:
+        print(f'close = {chart.closes[-1]} opens = {chart.opens[-1]} highs = {chart.highs[-1]} lows = {chart.lows[-1]}', file=sys.stderr)
         return Action_state.NEUTRAL
