+++ src/Chart.py
@@ -1,0 +1,26 @@
+##
+## EPITECH PROJECT, 2024
+## trading-bot
+## File description:
+## Chart
+##
+
+from Candle import Candle
+
+class Chart:
+    def __init__(self):
+        self.dates = []
+        self.opens = []
+        self.highs = []
+        self.lows = []
+        self.closes = []
+        self.volumes = []
+        self.indicators = {}
+
+    def add_candle(self, candle: Candle):
+        self.dates.append(candle.date)
+        self.opens.append(candle.open)
+        self.highs.append(candle.high)
+        self.lows.append(candle.low)
+        self.closes.append(candle.close)
+        self.volumes.append(candle.volume)
