+++ src/main.py
@@ -11,6 +11,8 @@ from BotAction import BotAction
 from ActionState import Action_state
 from Limit import Limit
 from RiskIndicator import RiskIndicator
+from BotState import BotState
+from CandlePatern import CandlePatern
 
 class Bot:
     def __init__(self):
@@ -21,6 +23,7 @@ class Bot:
         self.limit = Limit(1.2, 0.95)
         self.risk = RiskIndicator(13)
         self.bollinger = BollingerBands(13, 2)
+        self.candlePatern = CandlePatern()
 
     def run(self):
         while True:
@@ -50,8 +53,9 @@ class Bot:
 
             macd_state = self.macd.get_macd_state(affordable, bitcoin)
             rsi_state = self.rsi.get_rsi_state(affordable, bitcoin, self.botAction)
+            self.candlePatern.useCandlePatern(self.botState.charts["USDT_BTC"])
             bollinger_state = self.bollinger.get_bollinger_state(self.botState.closing_prices[-1])
-            print(f"{bollinger_state=}", file=sys.stderr)
+            # print(f"{bollinger_state=}", file=sys.stderr)
 
             # if self.limit.loss_limit(self.botState.closing_prices[-1]):
             #     if bitcoin > 0.001:
@@ -82,109 +86,6 @@ class Bot:
             #     self.limit.update_sell()
             #     return
             self.botAction.passAction()
-
-
-
-class Candle:
-    def __init__(self, format, intel):
-        tmp = intel.split(",")
-        for (i, key) in enumerate(format):
-            value = tmp[i]
-            if key == "pair":
-                self.pair = value
-            if key == "date":
-                self.date = int(value)
-            if key == "high":
-                self.high = float(value)
-            if key == "low":
-                self.low = float(value)
-            if key == "open":
-                self.open = float(value)
-            if key == "close":
-                self.close = float(value)
-            if key == "volume":
-                self.volume = float(value)
-
-    def __repr__(self):
-        return str(self.pair) + str(self.date) + str(self.close) + str(self.volume)
-
-
-class Chart:
-    def __init__(self):
-        self.dates = []
-        self.opens = []
-        self.highs = []
-        self.lows = []
-        self.closes = []
-        self.volumes = []
-        self.indicators = {}
-
-    def add_candle(self, candle: Candle):
-        self.dates.append(candle.date)
-        self.opens.append(candle.open)
-        self.highs.append(candle.high)
-        self.lows.append(candle.low)
-        self.closes.append(candle.close)
-        self.volumes.append(candle.volume)
-
-
-class BotState:
-    def __init__(self):
-        self.timeBank = 0
-        self.maxTimeBank = 0
-        self.timePerMove = 1
-        self.candleInterval = 1
-        self.candleFormat = []
-        self.candlesTotal = 0
-        self.candlesGiven = 0
-        self.initialStack = 0
-        self.transactionFee = 0.2
-        self.date = 0
-        self.stacks = dict()
-        self.charts = dict()
-        self.closing_prices = []
-
-    def update_chart(self, pair: str, new_candle_str: str):
-        if not (pair in self.charts):
-            self.charts[pair] = Chart()
-        new_candle_obj = Candle(self.candleFormat, new_candle_str)
-        self.charts[pair].add_candle(new_candle_obj)
-
-    def update_stack(self, key: str, value: float):
-        self.stacks[key] = value
-
-    def update_settings(self, key: str, value: str):
-        if key == "timebank":
-            self.maxTimeBank = int(value)
-            self.timeBank = int(value)
-        if key == "time_per_move":
-            self.timePerMove = int(value)
-        if key == "candle_interval":
-            self.candleInterval = int(value)
-        if key == "candle_format":
-            self.candleFormat = value.split(",")
-        if key == "candles_total":
-            self.candlesTotal = int(value)
-        if key == "candles_given":
-            self.candlesGiven = int(value)
-        if key == "initial_stack":
-            self.initialStack = 2404.97
-        if key == "transaction_fee_percent":
-            self.transactionFee = float(value)
-
-    def update_game(self, key: str, value: str):
-        if key == "next_candles":
-            new_candles = value.split(";")
-            self.date = int(new_candles[0].split(",")[1])
-            for candle_str in new_candles:
-                candle_infos = candle_str.strip().split(",")
-                self.update_chart(candle_infos[0], candle_str)
-        if key == "stacks":
-            new_stacks = value.split(",")
-            for stack_str in new_stacks:
-                stack_infos = stack_str.strip().split(":")
-                self.update_stack(stack_infos[0], float(stack_infos[1]))
-
 
 if __name__ == "__main__":
     mybot = Bot()
